refactor(flows): split the lint domain into one file per job
The fixes pushed both the module and its test past the line limits oxlint
enforces, so file selection and the team config lookup each move out on
their own. Also reformats the lint config the fixes edited, which was not
written the way oxfmt writes it.
